Add files via upload
This is just a rough draft to demonstrate that the files are working, the app file is a fake main meant to call the other files so it can be ignored. I am still working on the code for this, this is just to show what I have so far.